JSObject: Support full array indices range; use Longs for array indices
This commit supports the full extended range of array lengths.
Previously, the maximum array index allowed in an object's indexed
storage was 2^31-1, but it is not 2^32-1. We also now use Longs in many
places regarding arrays instead of Ints.

Unfortunately, as this is a large change, it introduces quite a few
regressions. These will hopefully be fixed in future commits

+++ src/main/kotlin/me/mattco/reeva/runtime/arrays/JSArrayObject.kt
@@ -24,7 +24,7 @@ open class JSArrayObject protected constructor(realm: Realm, proto: JSValue = re
     @JSNativePropertySetter("length", "ceW")
     fun setLength(thisValue: JSValue, newLength: JSValue): JSValue {
         expect(thisValue is JSObject)
-        thisValue.indexedProperties.setArrayLikeSize(Operations.toLength(newLength).asInt)
+        thisValue.indexedProperties.setArrayLikeSize(Operations.toLength(newLength).asLong)
         return JSUndefined
     }
 
@@ -68,7 +68,7 @@ open class JSArrayObject protected constructor(realm: Realm, proto: JSValue = re
             }.forEach {
                 val deleteSucceeded = delete(it)
                 if (!deleteSucceeded) {
-                    newLenDesc.setActualValue(this, (it + 1).toValue())
+                    newLenDesc.setActualValue(this, (it + 1L).toValue())
                     if (!newWritable)
                         newLenDesc.setWritable(false)
                     super.defineOwnProperty(property, newLenDesc)
